+++ CHANGELOG.md
@@ -6,6 +6,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add a new component `AdaptableGrid` to handle grids. From now, you can use this component instead of `Grid`.
 
+## [2.0.1] - 2019-04-11
+
+Fix:
+- Remove `getScssPaths` from `@kisskissbankbank/kitten`.
+  Prefer use `@kisskissbankbank/kitten/src/config/paths`.
+
+## [2.0.0] - 2019-04-10
+
+Breaking changes:
+- Upgrade React/ReactDOM dependencies (`>= 16.8.x`).
+- `getImagesPaths` has been removed and `getScssPaths` moved.
+  If you use Kitten paths, you will be update require by:
+  `import { getScssPaths } from '@kisskissbankbank/kitten'.
+
+Feature:
+- You can now import components from `@kisskissbankbank/kitten`.
+  Example: `import { Button, Text, Marger } from '@kisskissbankbank/kitten'.
+
 ## [1.26.0] - 2019-04-02
 
 Feature: